Add metrics to project

+++ app/initialization.py
@@ -3,7 +3,7 @@ import os
 import shutil
 import composer
 import configuration
-import security_checker
+import downloader
 
 
 def run():
@@ -12,16 +12,13 @@ def run():
 
     project_dir = sys.argv[1]
     build_dir = project_dir+'build/'
-    check_dir = build_dir+'check/'
 
     configuration.load(project_dir)
     configuration.add('project-dir', project_dir)
     configuration.add('build-dir', build_dir)
-    configuration.add('check-dir', check_dir)
 
     composer.initialization()
-    security_checker.initialization()
-    prepare_dir(check_dir)
+    downloader.initialization()
 
 
 def update():
@@ -31,9 +28,10 @@ def update():
     configuration.add('php', sys.argv[1])
     composer.initialization()
     composer.update()
-    security_checker.update()
+    downloader.update()
 
 
 def prepare_dir(path):
-    shutil.rmtree(path)
+    if os.path.isdir(path):
+        shutil.rmtree(path)
     os.makedirs(path)
